@@ -1,28 +1,45 @@
 {
     "scripts": {
-        "start": "node --inspect node_modules/webpack-dev-server/bin/webpack-dev-server.js",
-        "build": "webpack",
+        "start:client": "webpack serve",
+        "start:server": "nodemon server/index.ts",
+        "build": "webpack && tsc --project tsconfig.server.json",
         "stats": "webpack --profile --json > dist/stats.json"
     },
     "dependencies": {
+        "@types/express": "^4.17.11",
+        "@types/express-ws": "^3.0.0",
         "@types/lodash-es": "^4.17.3",
         "easyhard": "file:../build/core",
         "easyhard-api": "file:../build/api",
+        "easyhard-client": "file:../build/client",
+        "easyhard-common": "file:../build/common",
         "easyhard-loader": "file:../build/loader",
+        "easyhard-router": "file:../build/router",
+        "easyhard-server": "file:../build/server",
         "easyhard-styles": "file:../build/styles",
+        "express": "^4.17.1",
+        "express-ws": "^4.0.0",
+        "history": "^5.0.0",
         "jss": "^10.0.0",
         "jss-preset-default": "^10.0.0",
-        "lodash-es": "^4.17.15"
+        "lodash-es": "^4.17.15",
+        "module-alias": "^2.2.2",
+        "tsconfig-paths": "^3.9.0"
     },
     "devDependencies": {
         "babel-plugin-overload": "0.0.2",
         "eslint-loader": "^4.0.2",
-        "html-webpack-plugin": "^3.2.0",
+        "html-webpack-plugin": "4.5.0",
+        "nodemon": "^2.0.7",
         "recast": "^0.19.0",
-        "ts-loader": "^6.2.1",
-        "typescript": "3.3.3",
-        "webpack": "^4.41.5",
-        "webpack-cli": "^3.3.10",
-        "webpack-dev-server": "^3.9.0"
+        "ts-loader": "8.0.7",
+        "ts-node": "^9.1.1",
+        "typescript": "4.0.5",
+        "webpack": "^5.3.2",
+        "webpack-cli": "^4.1.0",
+        "webpack-dev-server": "3.11.0"
+    },
+    "_moduleAliases": {
+        "easyhard-common": "../build/common"
     }
 }
