@@ -2,7 +2,9 @@
   "scripts": {
     "build": "rollup -c",
     "build:dev": "rollup -cw",
-    "lint": "./node_modules/.bin/eslint packages/*/src/**/*",
+    "lint": "eslint packages/*/src/**",
+    "lint:examples": "eslint examples/src/**",
+    "test": "jest",
     "commit": "git-cz"
   },
   "repository": {
@@ -16,26 +18,36 @@
   },
   "homepage": "https://github.com/Ni55aN/easyhard.js#readme",
   "devDependencies": {
-    "@rollup/plugin-alias": "^3.1.1",
-    "@rollup/plugin-typescript": "^4.1.2",
-    "@typescript-eslint/eslint-plugin": "^2.12.0",
-    "@typescript-eslint/parser": "^2.12.0",
-    "commitizen": "^4.1.2",
-    "cz-conventional-changelog": "^3.2.0",
-    "eslint": "^6.7.2",
-    "eslint-loader": "^3.0.3",
+    "@rollup/plugin-typescript": "6.1.0",
+    "@testing-library/jest-dom": "^5.11.6",
+    "@types/jest": "^26.0.15",
+    "@types/ws": "^7.4.1",
+    "@typescript-eslint/eslint-plugin": "4.6.0",
+    "@typescript-eslint/parser": "4.6.0",
+    "commitizen": "4.2.2",
+    "core-js": "^3.7.0",
+    "cz-conventional-changelog": "3.3.0",
+    "eslint": "7.12.1",
+    "eslint-loader": "4.0.2",
+    "fast-check": "^2.7.0",
     "fromentries": "^1.2.1",
     "fs-extra": "^9.0.1",
+    "history": "^5.0.0",
     "husky": "^4.2.5",
-    "rollup": "^2.17.1",
+    "jest": "^26.6.3",
+    "recast": "^0.20.4",
+    "rollup": "2.32.1",
     "rollup-plugin-copy": "^3.3.0",
     "rollup-plugin-eslint": "^7.0.0",
-    "rollup-plugin-typescript2": "^0.27.1",
     "rxjs": "^7.0.0-beta.8",
-    "ts-loader": "^6.2.1",
-    "typescript": "^3.9.5",
     "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10"
+    "webpack-cli": "^3.3.10",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-typescript2": "0.29.0",
+    "ts-jest": "^26.4.4",
+    "ts-loader": "8.0.7",
+    "typescript": "4.0.5",
+    "ws": "^7.4.5"
   },
   "husky": {
     "hooks": {
