@@ -1,28 +1,22 @@
-import { Observable } from 'rxjs';
-import { first } from 'rxjs/operators';
-import { untilExist } from '../operators';
-import { $ } from '../structures/value';
+import { $ } from 'easyhard-common'
+import { first } from 'rxjs/operators'
+import { untilExist } from '../operators'
 
-type Cb = () => void
+type Callback = () => void
 
-const onMountElement = (el: Element, cb: Cb) => $('').pipe(untilExist(el), first()).subscribe(cb)
-const onMountInjected = (cb: Cb) => new Observable<null>(() => cb())
-
-export function onMount(...args: Parameters<typeof onMountElement>): void;
-export function onMount(...args: Parameters<typeof onMountInjected>): Observable<null>
-export function onMount(...args: Parameters<typeof onMountElement> | Parameters<typeof onMountInjected>) {
-  const [el, cb] = args
-  if (typeof el === 'function') return onMountInjected(el)
-  else onMountElement(el, cb as Cb)
+export function onMount(el: ChildNode, callback: Callback): void {
+  $('').pipe(untilExist(el), first()).subscribe({ next: callback })
 }
 
-const onDestroyElement = (el: Element, cb: Cb) => $('').pipe(untilExist(el)).subscribe({ complete: cb })
-const onDestroyInjected = (cb: Cb) => new Observable<null>(() => cb)
+export function onLife(el: ChildNode, callback: () => (void | Callback)): void {
+  let destroy: void | Callback
 
-export function onDestroy(...args: Parameters<typeof onDestroyElement>): void;
-export function onDestroy(...args: Parameters<typeof onDestroyInjected>): Observable<null>
-export function onDestroy(...args: Parameters<typeof onDestroyElement> | Parameters<typeof onDestroyInjected>) {
-  const [el, cb] = args
-  if (typeof el === 'function') return onDestroyInjected(el)
-  else onDestroyElement(el, cb as Cb)
+  $('').pipe(untilExist(el)).subscribe({
+    next() { destroy = callback() },
+    complete() { destroy && destroy() }
+  })
+}
+
+export function onDestroy(el: ChildNode, callback: Callback): void {
+  $('').pipe(untilExist(el)).subscribe({ complete: callback })
 }
