@@ -1,11 +1,12 @@
-export * from './core';
-export * from './types';
-export * from './operators';
-export * from './mutation-observer';
-export { createElement as h } from './core';
-export { createElementX as hx } from './jsx';
-export { $ } from './structures/value';
-export { $$ } from './structures/array';
-export * from './operators';
-export * from './utils';
-export * from './hooks';
+export { $, $$ } from 'easyhard-common'
+
+export * from './core'
+export * from './types'
+export * from './operators'
+export * from './mutation-observer'
+export { createElement as h } from './core'
+export { createElementX as hx } from './jsx'
+export * from './operators'
+export * from './utils'
+export * from './hooks'
+export * from './fragment'
