@@ -1,16 +1,17 @@
-import { Observable, MonoTypeOperatorFunction } from "rxjs";
-import { observeElement } from "../mutation-observer";
+import { Observable, MonoTypeOperatorFunction } from 'rxjs'
+import { observeElement } from '../mutation-observer'
 
 export function untilExist<T>(el: ChildNode | null, container: Node = document.body): MonoTypeOperatorFunction<T> {
   return <T>(source: Observable<T>): Observable<T> => new Observable(observer => {
-    const values: T[] = [];
+    const values: T[] = []
+    const next = () => {
+      while(values.length > 0) {
+        observer.next(values.shift())
+      }
+    }
 
     if (el) observeElement(el, {
-      added() {
-        while(values.length > 0) {
-          observer.next(values.shift())
-        }
-      },
+      added: next,
       removed() {
         observer.complete()
       }
@@ -18,14 +19,13 @@ export function untilExist<T>(el: ChildNode | null, container: Node = document.b
 
     return source.subscribe({
       next(value) {
-        values.push(value);
+        values.push(value)
         if (Boolean(el) && container.contains(el)) {
-          observer.next(values.pop());
-          values.splice(0, values.length)
+          next()
         }
       },
-      error(err) { observer.error(err); },
-      complete() { observer.complete(); }
-    });
-  });
+      error(err) { observer.error(err) },
+      complete() { observer.complete() }
+    })
+  })
 }
